Updated calculator reactivity
Changed to a system that won't just convert whatever is on screen to a javascript code expression. Instead opted to add a separate string that updates instead which is then used for the calculation. This allows for more user friendly input and screen reading.

Also tried to use floats wherever possible to avoid random rounding of answers.