Rewrote calculator reactivity, added math.js
Was having issues with the default way javascript handles floating points and ints and whatever. Implemented math.js to ease some of those issues because I can't be asked to write a whole system out for that at the moment.

Added extra buttons for inverse sin, cos and tan. Also added Euler's number, powered multiplication and factorials. Added a radians and degrees toggle but havent implemented it yet.

Code is still a bit messy and not written as well as it could be. Looking to improve that in the future.